# chore(dataobj): apply predicates sequentially