**What this PR does / why we need it**:

Updates the dataset reader to accept multiple predicates that get evaluated sequentially. When evaluating a predicate in the list, only columns that are required to evaluate it are loaded. Rows are filtered at each step of the evaluation with subsequent predicate evaluation having to fill the reduced row range.

For now the reader expects the caller to order predicates in a way to benefit from selectivity ~~and to ensure that a column is not referenced in multiple predicates from the list <- this might change in a follow-up commit or pr~~


```
goos: darwin
goarch: arm64
pkg: github.com/grafana/loki/v3/pkg/logql/bench
cpu: Apple M1 Pro
                                                                                                         │ /tmp/main.txt │            /tmp/seq.txt            │
                                                                                                         │    sec/op     │   sec/op     vs base               │
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_[BACKWARD]/kind=log/store=dataobj-8     545.4m ± 7%   498.3m ± 5%  -8.64% (p=0.000 n=30)

                                                                                                         │   /tmp/main.txt    │                /tmp/seq.txt                │
                                                                                                         │ kilobytesProcessed │ kilobytesProcessed  vs base                │
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_[BACKWARD]/kind=log/store=dataobj-8         102.78k ± 0%          54.28k ± 0%  -47.19% (p=0.000 n=30)
```


**Which issue(s) this PR fixes**:
Fixes #<issue number>

**Special notes for your reviewer**:

**Checklist**
- [ ] Reviewed the [`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md) guide (**required**)
- [ ] Documentation added
- [ ] Tests updated
- [ ] Title matches the required conventional commits format, see [here](https://www.conventionalcommits.org/en/v1.0.0/)
  - **Note** that Promtail is considered to be feature complete, and future development for logs collection will be in [Grafana Alloy](https://github.com/grafana/alloy). As such, `feat` PRs are unlikely to be accepted unless a case can be made for the feature actually being a bug fix to existing behavior.
- [ ] Changes that require user attention or interaction to upgrade are documented in `docs/sources/setup/upgrade/_index.md`
- [ ] If the change is deprecating or removing a configuration option, update the `deprecated-config.yaml` and `deleted-config.yaml` files respectively in the `tools/deprecated-config-checker` directory. [Example PR](https://github.com/grafana/loki/pull/10840/commits/0d4416a4b03739583349934b96f272fb4f685d15)