@ashwanthgoli 2025-04-25
updated the pr to add benchmarks for the reader. I have added one to measure gains from sequential predicate evaluation alone (not using page pruning)
benchmark columns here are:
- combined predicate
- sequential predicates with low selectivity first
- sequential predicates with high selectivity first

~5% reduction in cpu where both predicates are int64 columns. i think the gains depend a lot depends on data distribution, we could see better number from moving larger columns (log message for ex) to the right.

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/loki/v3/pkg/dataobj/internal/dataset
cpu: Apple M1 Pro
                     │  combined   │                low                 │                high                │
                     │   sec/op    │   sec/op     vs base               │   sec/op     vs base               │
PredicateExecution-8   97.65m ± 2%   98.32m ± 3%  +0.68% (p=0.043 n=10)   92.62m ± 1%  -5.16% (p=0.000 n=10)

                     │   combined   │                 low                  │                 high                 │
                     │     B/op     │     B/op      vs base                │     B/op      vs base                │
PredicateExecution-8   330.0Ki ± 0%   269.9Ki ± 0%  -18.21% (p=0.000 n=10)   241.9Ki ± 0%  -26.69% (p=0.000 n=10)

                     │  combined   │                 low                 │                high                │
                     │  allocs/op  │  allocs/op   vs base                │  allocs/op   vs base               │
PredicateExecution-8   11.82k ± 0%   13.82k ± 0%  +16.94% (p=0.000 n=10)   10.98k ± 0%  -7.10% (p=0.000 n=10)
```